Use atomic variable for num_tasks_completed, remove lock
Using atomic allow us the increment and comparison to be atomic, so we
can remove the lock. Overall, doesn't really affect performance.
More percisely, because in the main thread we weren't locking the run
reocrd when checking whether it's done (see the TODO in code), locking
here around `run_info->is_done` is not useful. Still suspecting there is
a hidden bug caused by this.